translate_api/requirements.txt
boto3
flask
python-dotenv
gunicorn